Move new preflight checks to preflight module
