+++ cumulusci/cumulusci.yml
@@ -35,12 +35,18 @@ tasks:
         class_path: cumulusci.tasks.apex.batch.BatchApexWait
         group: Salesforce
     check_communities_enabled:
-        description: Checks whether Communities are enabled.
-        class_path: cumulusci.tasks.salesforce.communities_preflights.IsCommunitiesEnabled
+        description: Runs as a preflight check to determine whether Communities are enabled.
+        class_path: cumulusci.tasks.preflights.sobjects.CheckSObjectsAvailable
+        options:
+            sobjects: Network
+        group: Salesforce Preflight Checks
+    check_sobjects_available:
+        description: Runs as a preflight check to determine whether specific sObjects are available.
+        class_path: cumulusci.tasks.preflights.sobjects.CheckSObjectsAvailable
         group: Salesforce Preflight Checks
     check_org_wide_defaults:
         description: Runs as a preflight check to validate Organization-Wide Defaults.
-        class_path: cumulusci.tasks.salesforce.sobject_preflights.CheckSobjectOWDs
+        class_path: cumulusci.tasks.salesforce.sobjects.CheckSobjectOWDs
         group: Salesforce Preflight Checks
     custom_settings_value_wait:
         description: Waits for a specific field value on the specified custom settings object and field
